zad4/zad4.1/python/requirements.txt
termcolor
grpcio
